Kernel/Net: Merge EtherType.h and EthernetFrameHeader.h
