+++ Kernel/Net/Ethernet.h
@@ -8,6 +8,17 @@
 
 #include <AK/Endian.h>
 #include <AK/MACAddress.h>
+#include <AK/Types.h>
+
+namespace Kernel {
+
+struct EtherType {
+    enum : u16 {
+        ARP = 0x0806,
+        IPv4 = 0x0800,
+        IPv6 = 0x86DD,
+    };
+};
 
 #pragma GCC diagnostic ignored "-Warray-bounds"
 
@@ -36,3 +47,4 @@ private:
 };
 
 static_assert(sizeof(EthernetFrameHeader) == 14);
+}
